Adjusts primitive base class and byteclass generics for better consistency during initialization.
